# src/main/java/com/ebuddy/controller/ContactController.java
package com.ebuddy.controller;
import com.ebuddy.service.ContactService; import com.ebuddy.web.ApiModels; import jakarta.validation.constraints.*; import org.springframework.http.*; import org.springframework.security.core.Authentication; import org.springframework.web.bind.annotation.*; import java.util.*;
@RestController @RequestMapping("/api/contacts") public class ContactController { final ContactService contacts; public ContactController(ContactService c){contacts=c;} private UUID u(Authentication a){return(UUID)a.getPrincipal();} @GetMapping public List<ApiModels.ContactView> list(Authentication a){return contacts.list(u(a));} @PostMapping @ResponseStatus(HttpStatus.CREATED) public ApiModels.ContactView add(Authentication a,@RequestParam @NotBlank String username){return contacts.add(u(a),username);} @PatchMapping("/{id}") public ApiModels.ContactView update(Authentication a,@PathVariable UUID id,@RequestParam String state,@RequestParam(required=false) String alias){return contacts.update(u(a),id,state,alias);} }